Upgraded to micronaut GA
