+++ .gitignore
@@ -2,3 +2,6 @@
 .gradle
 build
 out
+/bin
+/.classpath
+/.project
