+++ build.gradle
@@ -18,8 +18,8 @@ plugins {
     id "java-library"
 }
 
-group "be.looorent"
-version "1.3.0"
+group "uk.co.foonkiemonkey"
+version "1.0.0"
 description "Create API middleware to check Authorization headers against Keycloak"
 
 sourceCompatibility = 1.8
@@ -30,21 +30,21 @@ repositories {
     jcenter()
 }
 
-def keycloakVersion = "4.1.0.Final"
+def keycloakVersion = "4.5.0.Final"
 def rxJavaVersion = "2.1.16"
-def micronautVersion = "1.0.0.M4"
+def micronautVersion = "1.0.0"
 
 dependencies {
-    annotationProcessor "io.micronaut:inject-java:$micronautVersion"
+    annotationProcessor "io.micronaut:micronaut-inject-java:$micronautVersion"
 
     api "io.jsonwebtoken:jjwt:0.9.1"
 
-    implementation "io.micronaut:core:$micronautVersion"
-    implementation "io.micronaut:http:$micronautVersion"
-    implementation "io.micronaut:inject:$micronautVersion"
+    implementation "io.micronaut:micronaut-core:$micronautVersion"
+    implementation "io.micronaut:micronaut-http:$micronautVersion"
+    implementation "io.micronaut:micronaut-inject:$micronautVersion"
     implementation "io.reactivex.rxjava2:rxjava:$rxJavaVersion"
 
-    testImplementation "io.micronaut:http:$micronautVersion"
+    testImplementation "io.micronaut:micronaut-http:$micronautVersion"
     testImplementation "io.reactivex.rxjava2:rxjava:$rxJavaVersion"
     testImplementation "org.keycloak:keycloak-common:$keycloakVersion"
     testImplementation "org.keycloak:keycloak-core:$keycloakVersion"
@@ -71,9 +71,9 @@ apply plugin: "io.codearte.nexus-staging"
 publishing {
     publications {
         maven(MavenPublication) {
-            groupId "be.looorent"
+            groupId "uk.co.foonkiemonkey"
             artifactId "keycloak-micronaut-adapter"
-            version "1.3.0"
+            version "1.0.0"
             from components.java
         }
     }
@@ -125,12 +125,12 @@ if (project.hasProperty("release")) {
                         name project.name
                         description project.description
                         packaging "jar"
-                        url "https://github.com/looorent/keycloak-micronaut-adapter"
+                        url "https://github.com/jdariasl80/keycloak-micronaut-adapter"
 
                         scm {
-                            connection "scm:git:https://github.com/looorent/keycloak-micronaut-adapter"
+                            connection "scm:git:https://github.com/jdariasl80/keycloak-micronaut-adapter"
                             developerConnection "scm:git:git@github.com:looorent/keycloak-micronaut-adapter.git"
-                            url "https://github.com/looorent/keycloak-micronaut-adapter"
+                            url "https://github.com/jdariasl80/keycloak-micronaut-adapter"
                         }
 
                         licenses {
@@ -143,9 +143,9 @@ if (project.hasProperty("release")) {
 
                         developers {
                             developer {
-                                id = "looorent"
-                                name = "Lorent Lempereur"
-                                email = "lorent.lempereur.dev@gmail.com"
+                                id = "jdariasl80"
+                                name = "Jorge Arias"
+                                email = "jorge.arias@corestechno.com"
                             }
                         }
                     }
@@ -158,5 +158,5 @@ if (project.hasProperty("release")) {
 nexusStaging {
     username = repoUser
     password = repoPassword
-    packageGroup = "be.looorent"
+    packageGroup = "uk.co.foonkiemonkey"
 }
