--- src/main/java/be/looorent/micronaut/security/Constant.java
@@ -1,13 +1,0 @@
-package be.looorent.micronaut.security;
-
-/**
- * @author Lorent Lempereur - lorent.lempereur.dev@gmail.com
- */
-public class Constant {
-
-    /**
-     * Name of the HTTP Request's attribute set by this Keycloak adapter.
-     */
-    public static final String SECURITY_CONTEXT = "securityContext";
-
-}
