--- src/main/java/be/looorent/micronaut/security/SecurityContext.java
@@ -1,7 +1,0 @@
-package be.looorent.micronaut.security;
-
-/**
- * A data container describing the current security context.
- * @author Lorent Lempereur - lorent.lempereur.dev@gmail.com
- */
-public interface SecurityContext {}
