+++ src/main/java/uk/co/foonkiemonkey/micronaut/security/Constant.java
@@ -1,0 +1,13 @@
+package uk.co.foonkiemonkey.micronaut.security;
+
+/**
+ * @author Lorent Lempereur - lorent.lempereur.dev@gmail.com
+ */
+public class Constant {
+
+    /**
+     * Name of the HTTP Request's attribute set by this Keycloak adapter.
+     */
+    public static final String SECURITY_CONTEXT = "securityContext";
+
+}
