+++ src/main/java/uk/co/foonkiemonkey/micronaut/security/SecurityContext.java
@@ -1,0 +1,7 @@
+package uk.co.foonkiemonkey.micronaut.security;
+
+/**
+ * A data container describing the current security context.
+ * @author Lorent Lempereur - lorent.lempereur.dev@gmail.com
+ */
+public interface SecurityContext {}
