+++ src/main/java/uk/co/foonkiemonkey/micronaut/security/SecurityContextFactory.java
@@ -1,0 +1,17 @@
+package uk.co.foonkiemonkey.micronaut.security;
+
+import io.jsonwebtoken.Claims;
+
+/**
+ * Contract that can be implemented by the adapter's user to build their {@link SecurityContext}.
+ * @author Lorent Lempereur - lorent.lempereur.dev@gmail.com
+ */
+public interface SecurityContextFactory {
+
+    /**
+     * Starting from a JWT's body, builds a context of security.
+     * @param tokenContent a JWT's body; must not be null
+     * @return the built security context
+     */
+    SecurityContext createSecurityContext(Claims tokenContent);
+}
