+++ src/test/groovy/uk/co/foonkiemonkey/micronaut/security/SubjectSecurityContext.groovy
@@ -1,0 +1,13 @@
+package uk.co.foonkiemonkey.micronaut.security
+
+import groovy.transform.EqualsAndHashCode
+
+@EqualsAndHashCode(includes = ["subject"])
+class SubjectSecurityContext implements SecurityContext {
+
+    String subject
+
+    SubjectSecurityContext(String subject) {
+        this.subject = subject
+    }
+}
